pyproject.toml
[project]
name = "feedforbot"
version = "0.1.0"
description = "Forward links from RSS/Atom feeds to messengers"
authors = [{ name = "Aleksandr Shpak", email = "shpaker@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "croniter>=6.0",
  "beautifulsoup4>=4.13,<5",
  "feedparser>=6.0,<7",
  "httpx>=0.28",
  "jinja2>=3.1,<4",
  "pydantic>=2.0,<3",
  "epyxid>=0.3",
]

[project.optional-dependencies]
cli = [
  "click>=8.1,<9",
  "pyyaml>=6.0,<7",
  "redis>=5,<6",
  "sentry-sdk>=2.0,<3",
  "structlog>=24.1",
]

[project.scripts]
feedforbot = "feedforbot.cli:main"

[project.urls]
Homepage = "https://github.com/shpaker/feedforbot"
Repository = "https://github.com/shpaker/feedforbot"

[dependency-groups]
dev = [
  "feedforbot[cli]",
  "fakeredis>=2,<3",
  "freezegun>=1.5",
  "mypy>=1.15,<2",
  "pytest>=8.3",
  "pytest-asyncio>=0.25",
  "pytest-cov>=6.0",
  "respx>=0.22",
  "ruff>=0.11",
  "types-beautifulsoup4>=4.12",
  "types-croniter>=6.0",
  "types-pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.10"
plugins = [
  "pydantic.mypy",
]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
addopts = "--cov=feedforbot --cov-report=term-missing"

[[tool.mypy.overrides]]
module = ["feedparser.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 79
exclude = [
  ".venv",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
ignore = [
  "PLR0913",  # Too many arguments to function call
  "PTH123",   # `open("foo")` should be replaced by `Path("foo").open()`
  "RUF001",   # Docstring contains ambiguous chars
  "RUF002",   # Docstring contains ambiguous chars
  "RUF003",   # Docstring contains ambiguous chars
  "COM812",   # Trailing comma missing (conflicts with ruff format)
  "ISC001",   # Implicit string concatenation (conflicts with ruff format)
]
per-file-ignores."tests/**" = [
  "S101",   # assert is fine in tests
  "S105",   # hardcoded passwords (fake tokens in fixtures)
  "TRY003", # long exception messages in test helpers
]
flake8-tidy-imports.ban-relative-imports = "all"
mccabe.max-complexity = 20
isort.force-single-line = false
isort.lines-after-imports = 2
isort.known-first-party = ["feedforbot"]
select = [
   "F",  # Pyflakes
   "E",  # pycodestyle
 "C90",  # mccabe
   "I",  # isort
  "UP",  # pyupgrade
 "YTT",  # flake8-2020
   "S",  # flake8-bandit
 "BLE",  # flake8-blind-except
   "B",  # flake8-bugbear
 "COM",  # flake8-commas
  "C4",  # flake8-comprehensions
 "DTZ",  # flake8-datetimez
 "T10",  # flake8-debugger
 "EXE",  # flake8-executable
 "ISC",  # flake8-implicit-str-concat
   "G",  # flake8-logging-format
 "INP",  # flake8-no-pep420
 "T20",  # flake8-print
 "RET",  # flake8-return
 "SIM",  # flake8-simplify
 "TID",  # flake8-tidy-imports
 "TCH",  # flake8-type-checking
 "ARG",  # flake8-unused-arguments
 "PTH",  # flake8-use-pathlib
 "ERA",  # eradicate
  "PL",  # Pylint
 "TRY",  # tryceratops
 "RSE",  # flake8-raise
 "SLF",  # flake8-self
 "RUF",  # Ruff-specific rules
]
